works for 7 users, 50+ matches
Former-commit-id: c1d66853fec76f18c9d716f6ba07ea4e3678e9c1

+++ MatchService.js
@@ -17,9 +17,6 @@ var steam = new Steam(
     process.env.STEAM_GUARD_CODE);
 
 var aq = async.queue(apiRequest, 1)
-aq.empty = function(){
-    getNewGames();
-}
 var pq = async.queue(parseReplay, 1)
 var api_delay = 1000
 var api_url = "https://api.steampowered.com/IDOTA2Match_570/"
@@ -30,17 +27,24 @@ var parser_file = process.env.PARSER_FILE || "./parser/target/stats-0.1.0.jar"
 if (!fs.existsSync(replay_dir)){
     fs.mkdir(replay_dir)
 }
-//if env variable set, reparse all matches
+//reparse all matches
 if (process.env.RESET_ON_START){
-    console.log("[RESET] resetting parse status")
     matches.update({}, {$set:{parse_status: 0}}, { multi: true })
 }
-//put any unparsed matches onto parse queue
-matches.find({parse_status : 0}, function(err, docs) {
-    pq.push(docs, function (err){})
-})
 getNewGames()
+parseMatches()
 
+aq.empty = function(){
+    getNewGames();
+}
+pq.empty = function(){
+    parseMatches();
+}
+function parseMatches(){
+    matches.find({parse_status : 0}, function(err, docs) {
+        pq.push(docs, function (err){})
+    })
+}
 function getNewGames(){
     players.find({}, function(err,docs){
         aq.push(docs, function(err){})
@@ -54,13 +58,8 @@ function apiRequest(req, cb){
         data = data.result
         if (req.player_id){
             console.log("[API] games for player %s", req.player_id)
-            data.matches.forEach(function(match){
-                matches.findOne({ match_id: match.match_id }, function(err, doc) {
-                    if (err) {return cb(err)}
-                    if(!doc){
-                        aq.push(match, function(err){})
-                    }
-                })
+            async.mapSeries(data.matches, queueMatchDetails, function(err){
+                setTimeout(cb, api_delay, null)
             })
         }
         if (req.match_id){
@@ -68,72 +67,112 @@ function apiRequest(req, cb){
             data.parse_status = 0;
             matches.insert(data)
             pq.push(data, function(err){})
+            setTimeout(cb, api_delay, null)
         }
-        setTimeout(cb, api_delay, null)
     })
 }
 
+function queueMatchDetails(match, cb){
+    matches.findOne({ match_id: match.match_id }, function(err, doc) {
+        if(!doc){aq.push(match, function(err){})}
+        cb(null)
+    })
+}
 function download(match, cb) {
     var match_id = match.match_id
     var fileName = replay_dir+match_id+".dem"
+
     if (fs.existsSync(fileName)){
         console.log("[DL] found local replay for match %s", match_id)
-        return cb(null, fileName);
-    }
-    if (process.env.AWS_S3_BUCKET){
-        var s3 = new AWS.S3()
-        var params = {Bucket: process.env.AWS_S3_BUCKET, Key: fileName}
-        s3.getObject(params, function(err, data) {
-            if (!err){
-                console.log("[DL] found S3 replay for match %s", match_id)
-                fs.writeFileSync(fileName, data.Body);
-                return cb(null, fileName)
-            }
-        })
-    }
-    if (match.start_time > moment().subtract(8, 'days').format('X')){
-        getReplayUrl(match, function(url){
-            downloadWithRetry(url, fileName, 1000, function(){
-                console.log("[DL] found steam replay for match %s", match_id)
-                return cb(null, fileName)
-            })
-        })
+        cb(null, fileName);
     }
     else{
-        console.log("[DL] replay expired for match %s", match_id);
-        matches.update({match_id: match_id}, {$set: {parse_status : 1}})
-        return cb(true)
+        if (match.start_time > moment().subtract(7, 'days').format('X')) {
+            getReplayUrl(match, function(url){
+                downloadFromSteam(url, fileName, 1000, function(){
+                    console.log("[DL] found steam replay for match %s", match_id)
+                    uploadToS3(fileName, function(err){
+                        cb(null, fileName)
+                    })
+                })
+            })
+        }
+        else{
+            var s3 = new AWS.S3()
+            var params = {Bucket: process.env.AWS_S3_BUCKET, Key: fileName}
+            s3.getObject(params, function(err, data) {
+                if (!process.env.AWS_S3_BUCKET || err){
+                    console.log("[DL] replay expired for match %s", match_id)
+                    matches.update({match_id: match_id}, {$set: {parse_status : 1}})
+                    cb(true)
+                }
+                else{
+                    console.log("[DL] found S3 replay for match %s", match_id)
+                    fs.writeFileSync(fileName, data.Body);
+                    cb(null, fileName)
+                }
+            })
+        }
+    }
+}
+function uploadToS3(fileName, cb){
+    var s3 = new AWS.S3()
+    var params = { Bucket: process.env.AWS_S3_BUCKET, Key: fileName}
+    s3.headObject(params, function(err, data){
+        if (err){
+            params.Body = fs.readFileSync(fileName)
+            s3.putObject(params,function (err, data) {
+                if (!process.env.AWS_S3_BUCKET || err){
+                    console.log('[S3] could not upload to S3')
+                    cb(true);
+                }
+                else{
+                    console.log('[S3] Successfully uploaded replay to S3: %s ', fileName)
+                    cb(null)
+                }
+            })
+        }
+        else{
+            console.log('[S3] replay already exists in S3')
+            cb(null)
+        }
+    })
+}
+function getReplayUrl(match, cb){
+    if (match.replay_url) { 
+        console.log("[STEAM] found replay_url in db")
+        return cb(match.replay_url)
+    }
+    if (!steam.ready){
+        console.log("[STEAM] not ready yet")
+        setTimeout(getReplayUrl, 5000, match, cb)
+    }
+    else{
+        steam.getReplayDetails(match.match_id, function(err, data) {
+            if (err) {
+                //todo login with another account and try again
+                console.log(err)
+                setTimeout(getReplayUrl, 5000, match, cb)
+            }
+            else{
+                var url = "http://replay"+data.cluster+".valve.net/570/"+data.id+"_"+data.salt+".dem.bz2";
+                matches.update({match_id: match.match_id}, {$set: {replay_url:url}})
+                cb(url)
+            }
+        })   
     }
 }
 
-function getReplayUrl(match, cb){
-    if (match.replay_url) { return cb(match.replay_url)}
-    steam.getReplayDetails(match.match_id, function(err, data) {
-        if (err) {
-            console.log("[STEAM] steam not ready, retrying")
-            setTimeout(getReplayUrl, 5000, match, cb)
-        }
-        else{
-            var url = "http://replay"+data.cluster+".valve.net/570/"+data.id+"_"+data.salt+".dem.bz2";
-            matches.update({match_id: match.match_id}, {$set: {replay_url:url}})
-            cb(url)
-        }
-    })  
-}
-
-function downloadWithRetry(url, fileName, timeout, cb){
-    console.log('[DL] Downloading replay from %s', url)
-    request({url:url, encoding:null}, function (error, response, body) {
-        if (response.statusCode !== 200 || error) {
+function downloadFromSteam(url, fileName, timeout, cb){
+    request({url:url, encoding:null}, function (err, response, body) {
+        if (err || response.statusCode !== 200) {
             console.log("[DL] failed to download from %s, retrying in %ds", url, timeout/1000)
-            setTimeout(downloadWithRetry, timeout, url, fileName, timeout*2, cb);
+            setTimeout(downloadFromSteam, timeout, url, fileName, timeout*2, cb);
         }
         else{
-            console.log("[DL] decompressing replay to %s", fileName)
             body = Bunzip.decode(body);  
-            console.log("[DL] saving replay to %s", fileName)
             fs.writeFileSync(fileName, body);
-            cb()
+            cb(null)
         }
     })
 }
@@ -142,10 +181,11 @@ function parseReplay(match, cb){
     var match_id = match.match_id
     download(match, function(err, fileName){
         if (err) {
+            console.log("[DL] error downloading replay for match %s", match_id)
             return cb(err)
         }
-        console.log("[PARSER] started on %s", fileName);
-        var output="";
+        console.log("[PARSER] started on %s", fileName)
+        var output=""
         var cp = spawn(
             "java",
             ["-jar",
@@ -163,24 +203,9 @@ function parseReplay(match, cb){
             if (!code){
                 matches.update({match_id: match_id}, {$set: JSON.parse(output)})
                 matches.update({match_id: match_id}, {$set: {parse_status : 2}})
-                if (process.env.AWS_S3_BUCKET){
-                    var s3 = new AWS.S3()
-                    var params = { Bucket: process.env.AWS_S3_BUCKET, Key: fileName, Body: fs.readFileSync(fileName)}
-                    s3.putObject(params,function (err, data) {
-                        if (!err){
-                            console.log('[S3] Successfully uploaded replay to S3: %s ', fileName)
-                            if (process.env.DELETE_REPLAY_FILES){
-                                console.log("[DELETE] deleting replay: %s", fileName)
-                                fs.unlink(fileName)
-                            }
-                            cb(code)
-                        }
-                    })
-                }
+                if (process.env.DELETE_REPLAYS){fs.unlink(fileName)}
             }
-            else{
-                cb(code)
-            }
+            cb(code)
         })
     })
 }
